requirements.txt
pytest
Pandas
NumPy
setuptools
pyarrow
pyyaml
fastparquet